+++ component_catalog/models.py
@@ -1841,10 +1841,9 @@ class Package(
     def identifier(self):
         """
         Provide a unique value to identify each Package.
-        It is the Package URL (minus the 'pkg:' prefix) if one exists;
-        otherwise it is the Package Filename.
+        It is the Package URL if one exists; otherwise it is the Package Filename.
         """
-        return self.short_package_url or self.filename
+        return self.plain_package_url or self.filename
 
     @classmethod
     def identifier_help(cls):
@@ -1875,8 +1874,8 @@ class Package(
         Return the Package URL string as a valid filename.
         Useful when `Package.filename` is not available.
         """
-        cleaned_package_url = self.short_package_url
-        for char in "/@?=#":
+        cleaned_package_url = self.plain_package_url
+        for char in "/@?=#:":
             cleaned_package_url = cleaned_package_url.replace(char, "_")
         return get_valid_filename(cleaned_package_url)
 
@@ -2017,9 +2016,15 @@ class Package(
 
     @cached_property
     def component(self):
-        """Return the Component instance if 1 and only 1 Component is assigned to this Package."""
-        with suppress(ObjectDoesNotExist, MultipleObjectsReturned):
-            return self.component_set.get()
+        """
+        Return the Component instance if 1 and only 1 Component is assigned to this
+        Package.
+        Using ``component_set.all()`` to benefit from prefetch_related when it was
+        applied to the Package QuerySet.
+        """
+        component_set = self.component_set.all()
+        if len(component_set) == 1:
+            return component_set[0]
 
     def set_values_from_component(self, component, user):
         changed_fields = set_fields_from_object(